+++ src/main/kotlin/blackjack/model/winning/Betting.kt
@@ -1,3 +1,3 @@
-package blackjack.model
+package blackjack.model.winning
 
 data class Betting(val amount: Int)
